requirements.txt
aiosqlite==0.21.0
databases==0.9.0
greenlet==3.2.4
SQLAlchemy==2.0.43
typing_extensions==4.15.0
PyPDF2
sentence-transformers
langchain
pypdf
python-docx
fastapi
uvicorn
python-multipart